corosync: Add attribute to not start corosync on boot after fencing
If the node[:corosync][:check_for_start] attribute is set to true, then
corosync will not start on boot automatically, nor through
chef-client, if the file /var/cache/corosync/block_automatic_start
exists.

On proper shutdown, one new init script (corosync-clean-fencing) is used
to remove that file, resulting in the next boot to have corosync
starting.

This allows to not start corosync after fencing, to avoid a never-ending
deathmatch, or a fencing loop.